requirements.txt
psycopg2-binary
flask~=2.2.2
flask-cors
Flask-SQLAlchemy
flask-jwt-extended
SQLAlchemy~=1.4.46
pyjwt~=2.6.0
uuid~=1.30
python-dateutil~=2.8.2
uuid7
Werkzeug~=2.2.2